Update map ids from shorts to ints in a few remaining places

+++ CompatibilityLib/base/src/main/java/com/elmakers/mine/bukkit/utility/platform/base/DeprecatedUtilsBase.java
@@ -75,7 +75,7 @@ public abstract class DeprecatedUtilsBase implements DeprecatedUtils {
     }
 
     @Override
-    public short getMapId(MapView mapView) {
+    public int getMapId(MapView mapView) {
         return mapView.getId();
     }
