+++ CompatibilityLib/common/src/main/java/com/elmakers/mine/bukkit/utility/platform/DeprecatedUtils.java
@@ -37,7 +37,7 @@ public interface DeprecatedUtils {
 
     MapView getMap(int id);
 
-    short getMapId(MapView mapView);
+    int getMapId(MapView mapView);
 
     String getName(EntityType entityType);
